Partial refactor of rom_tools. Fixes: 	- Address translation works (or at least is compatible with SMILE's translation 	- RomManager now has control of two rom files - the "clean" rom and the "new" rom. The clean rom is never written to so that it does not need to be backed up.
Other changes:
	- Style changes in most of rom_tools to match the code style in the rest of the codebase
	- Added a FutureAddress to represent and address which is unknown. A FutureAddress uses an
environment to map itself to a real pointer. Allocation produces FutureAddressWrites which are places
on the ROM where an unknown pointer will be written. Allocation also produces new known pointers,
and fills the env with them. Once everything is allocated, the FutureAddressWrites can be resolved
to actual bytes on the ROM.
	- Started fleshing out the room data structures in level.py to account for new objects like
room state, fx, scrolls, etc. that weren't considered in the first draft.