Fix RNG for concrete map generation
Eliminates the use of `set()` in several crucial locations. The order of
iteration over a `set()` is not guaranteed, while the order of a `dict`
is guaranteed by Python 3.7+. By replacing sets with dicts, this change
means that the same concrete map is generated for the same seed. Other
parts of the codebase may need to receive similar scrutiny.